requirements.txt
asgiref==3.5.0
Django==4.0.3
django-filter==21.1
django-simple-history==3.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
